Add dh_eoscontent for merging CMS content to app packages
Various chunks of data provided by the Endless CMS are meant to override
the data that comes from the debian package. Currently that involves
installing extra files that override the data at runtime, but it would
be better if the CMS data was simply included in the package.

dh_eoscontent reads the data installed by eos-shell-content and merges
it into the package. Currently this just handles desktop files.